fix(sql): remove pivot query import cycle
Define PivotQuery after Query in the same module to prevent Turbopack production bundles from reading the superclass before initialization. Preserve the public exports and factory identity, and remove the obsolete pivot module.
